Move common components into shared directory.
