+++ myuw/static/vue/components/shared/cards/husky-experience.vue
@@ -41,7 +41,7 @@
 
 <script>
 import {mapGetters, mapState, mapActions} from 'vuex';
-import Card from '../layouts/card.vue';
+import Card from '../../layouts/card.vue';
 
 export default {
   components: {
@@ -107,7 +107,7 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-@import "../../../css/myuw/variables.scss";
+@import '../../../../css/myuw/variables.scss';
 
 .myuw-huskyexp {
   a {
