+++ myuw/static/vue/tests/outage.test.js
@@ -1,7 +1,7 @@
 import {shallowMount} from '@vue/test-utils';
 import Vuex from 'vuex';
 import {createLocalVue} from './helper';
-import Outage from '../components/cards/outage.vue';
+import Outage from '../components/shared/cards/outage.vue';
 
 import {library} from '@fortawesome/fontawesome-svg-core';
 import {
