+++ myuw/static/vue/tests/visual-schedule.test.js
@@ -8,9 +8,9 @@ import schedule from '../vuex/store/schedule';
 import {statusOptions} from '../vuex/store/model_builder';
 import {expectAction} from './helper';
 
-import CourseSection from '../components/cards/schedule/course-section.vue';
-import ScheduleTab from '../components/cards/schedule/schedule-tab.vue';
-import VisualSchedule from '../components/cards/schedule/visual-schedule.vue';
+import CourseSection from '../components/shared/cards/schedule/course-section.vue';
+import ScheduleTab from '../components/shared/cards/schedule/schedule-tab.vue';
+import VisualSchedule from '../components/shared/cards/schedule/visual-schedule.vue';
 
 import mockScheduleBill from './mock_data/schedule/bill2013.json';
 import mockScheduleBillsea2020 from './mock_data/schedule/billsea2020.json';
